Fixed segfaults & crashes up til initialization of an opengl context.
This is where it will get tedious. I will have to replace GLX function
loading with EGL function loading. Gamma ramp has been disabled. A
bunch. It tricks the higher level APIs into thinking it's in XF86 mode. I chose that because it seems to have less checks to get to the next native function. AKA I can feed it garbage and it should work.